requirements.txt
beautifulsoup4==4.8.0
joblib==0.14.1
lxml==4.4.2
nltk==3.4.5
numpy==1.17.2
pandas==0.25.1
Pillow==6.2.1
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.2
scikit-learn==0.22
scipy==1.3.3
six==1.12.0
soupsieve==1.9.4
